pw: data: Check `data` field for NULL and return as an `Option`.
The `data` field of a `spa_data` may be NULL, so creating a slice from it is UB in those cases.
This makes the method return the slice in an `Option`, with `None` being returned in case it is NULL.

This also changes the name of the method to reflect the name of the accessed field.