libspa: Remove warning for missing docs
This commit disables warnings for undocumented items in the libspa crate.

For some items, like self-describing variants of an enum, adding documentation is just not worth it.
The fact that any warning here makes the whole CI pipeline appear as failed is also annoying when working on any merge requests.