pw: Replace `signal` dependency with `nix` dependency
The `signal` crate is not widely used and has not been updated for a while,
and depended on an outdated version of the `nix` crate.

The `nix` crate provides the same enum we needed from the `signal` crate
and is more actively maintained, so we can use it directly instead.