libspa-sys: Add "re-exports" for type-info arrays and spa_debug_type_find
This adds a new type-info.c file in libspa-sys which has redefinitions of type-info related symbols in the spa headers.
As SPA is a header-only library, the symbols in the SPA headers are marked `static` and can not be linked to from rust.

Therefore, we instead compile our type-info.c file into a C object and can then link to the redefinitions from rust.

This mechanism can be used to use any global variables, functions and (in a limited way) macros from SPA in the future.